third_party/blink/web_tests/fast/events/onerror-no-constructor-expected.txt

CONSOLE MESSAGE: FAIL
CONSOLE ERROR: Uncaught PASS if this is reported as an uncaught exception.
